#554: move dictionary to src/zmm
