+++ src/content_manager.h
@@ -39,7 +39,7 @@
 #include "autoscan.h"
 #include "cds_objects.h"
 #include "common.h"
-#include "dictionary.h"
+#include "zmm/dictionary.h"
 #include "generic_task.h"
 #include "storage.h"
 #include "timer.h"
