+++ src/file_request_handler.h
@@ -33,7 +33,7 @@
 #define __FILE_REQUEST_HANDLER_H__
 
 #include "common.h"
-#include "dictionary.h"
+#include "zmm/dictionary.h"
 #include "request_handler.h"
 #include "upnp_xml.h"
