+++ src/url_request_handler.cc
@@ -39,7 +39,7 @@
 
 #include "iohandler/buffered_io_handler.h"
 #include "cds_objects.h"
-#include "dictionary.h"
+#include "zmm/dictionary.h"
 #include "play_hook.h"
 #include "url_request_handler.h"
