+++ src/web/edit_save.cc
@@ -30,7 +30,7 @@
 /// \file edit_save.cc
 
 #include "content_manager.h"
-#include "dictionary.h"
+#include "zmm/dictionary.h"
 #include "pages.h"
 #include "storage.h"
 #include "tools.h"
