+++ src/zmm/dictionary.cc
