+++ test/test_runtime/test_runtime.h
@@ -22,7 +22,7 @@
 #ifndef __RUNTIME_TEST_H__
 #define __RUNTIME_TEST_H__
 
-#include "dictionary.h"
+#include "zmm/dictionary.h"
 #include "gtest/gtest.h"
 
 class RuntimeTest : public ::testing::Test {
